Fixed Pathfinding
Creatures now follow where they last saw the player if they lose sight
of them.

Also fixed getting stuck on walls, used to be a problem because they
couldnt see the player and the pathfinder wouldnt find them. Running
into walls due to wandering, not due to the pathfinder. This is fixed
now