Item UI Update
First of all, set up armor to work properly between Light, Medium and
Heavy. Requiring a strength base or a perk to not have a negative
modifier to your evasion, speed and attack.

Changed the inspect screen. Inspecting an item displays a short
description, a list of the ways it affects your stats, and a list of all
ItemTags on it (along with their descriptions)