Unoptimized Rest Function
Screen contains a boolean, if the boolean is true it is set to false and
the key press repeats automatically. Hitting 5 sets the player to
resting, and repeatedly hits 5 over and over until the rest is complete.
Unoptimized method of scanning the map to see if there are enemies in
sight to stop the rest