Lizardfolk Added
A bunch more enemies in the form of lizardfolk added. The casters do not
have valid spells currently and there is a high difficulty spike due to
lack of balance.

Reduced the amount of xp the player needs to level up from
level^(1.2)*750 to level^(1.2)*500 to hopefully help with balance.