Ability Framework + Rings
Abilities are granted (so far not at all) by equipment or perks. Staff
replaced by the spear which has the same base stats as the hand axe but
lets you reach 1 space away.  Abilities are accessed by [a]

Also did a bunch of rings. Rings have different art and descriptions
every game. Currently there is a ring of poison resistance, and a ring
for each stat and attribute that increases those things based on what
level they are found on