Music and Prefabs
Music has been added. Currently has 5 tracks that play at different
times, 4 of them overlap and fade in and out of each other depending on
situation (combat is more intense than main, so when there is a creature
in view it plays combat music)

Prefabs have been fixed. Chasms now work to allow for more interesting
prefabs. Set it up so that it does not delete dead ends that are part of
rooms. This also allows more interesting things in the future as dungeon
gen keeps track of what areas are rooms and what areas are not.