Basic Mouse Controls
Redid all of respondToUserInput to take a KeyCode, char and boolean
(shift) rather than a keyevent as this lets me simulate keypresses.
Currently it seems to work perfectly fine.
Mousing over things is annoying af but works. Added some more inputs and
events to main to support mouse controls.

So far so good for a wednesday release of demo 1